+++ admin/marker/extramarker.php
@@ -125,6 +125,8 @@ function leafext_extramarker_help() {
 	if ( is_singular() || is_archive() ) {
 		return $text;
 	} else {
-		echo wp_kses_post( $text );
+		// echo wp_kses_post( $text );
+		// phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
+		echo $text;
 	}
 }